chore: refactor gql queries
